+++ src/hessian_long.cc
@@ -23,7 +23,7 @@ short hessian_decode_long(uint8_t * const buf, const size_t buf_length, const v8
 		return 1;
 	}
 	if (buf_length >= 4 && code == 0x59) {
-		args.GetReturnValue().Set(ntohl(*(int32_t *)(buf+1)));
+		args.GetReturnValue().Set(static_cast<uint32_t>(ntohl(*(int32_t *)(buf+1))));
 		return 1;
 	}
 	if (buf_length >= 5 && code == 0x4c) {
